+++ docassemble_webapp/docassemble/webapp/core/models.py
@@ -7,8 +7,8 @@ class Uploads(db.Model):
     __tablename__ = dbtableprefix + "uploads"
     indexno = db.Column(db.Integer(), primary_key=True)
     key = db.Column(db.String(250), index=True)
-    filename = db.Column(db.Text(), index=True)
-    yamlfile = db.Column(db.Text())
+    filename = db.Column(db.String(255), index=True)
+    yamlfile = db.Column(db.String(255), index=True)
     private = db.Column(db.Boolean(), nullable=False, server_default=true())
     persistent = db.Column(db.Boolean(), nullable=False, server_default=false())
 
@@ -28,13 +28,13 @@ class UploadsRoleAuth(db.Model):
 class ObjectStorage(db.Model):
     __tablename__ = dbtableprefix + "objectstorage"
     id = db.Column(db.Integer(), primary_key=True)
-    key = db.Column(db.Text(), index=True)
+    key = db.Column(db.String(1024), index=True)
     value = db.Column(db.Text())
 
 class SpeakList(db.Model):
     __tablename__ = dbtableprefix + "speaklist"
     id = db.Column(db.Integer(), primary_key=True)
-    filename = db.Column(db.Text(), index=True)
+    filename = db.Column(db.String(255), index=True)
     key = db.Column(db.String(250), index=True)
     phrase = db.Column(db.Text())
     question = db.Column(db.Integer())
@@ -56,8 +56,8 @@ class Supervisors(db.Model):
 class MachineLearning(db.Model):
     __tablename__ = dbtableprefix + "machinelearning"
     id = db.Column(db.Integer(), primary_key=True)
-    group_id = db.Column(db.Text())
-    key = db.Column(db.Text(), index=True)
+    group_id = db.Column(db.String(1024))
+    key = db.Column(db.String(1024), index=True)
     independent = db.Column(db.Text())
     dependent = db.Column(db.Text())
     info = db.Column(db.Text())
@@ -69,11 +69,11 @@ class Shortener(db.Model):
     __tablename__ = dbtableprefix + "shortener"
     id = db.Column(db.Integer(), primary_key=True)
     short = db.Column(db.String(250), nullable=False, unique=True)
-    filename = db.Column(db.Text(), index=True)
+    filename = db.Column(db.String(255), index=True)
     uid = db.Column(db.String(250))
     user_id = db.Column(db.Integer(), db.ForeignKey(dbtableprefix + 'user.id', ondelete='CASCADE'))
     temp_user_id = db.Column(db.Integer(), db.ForeignKey(dbtableprefix + 'tempuser.id', ondelete='CASCADE'))
-    key = db.Column(db.Text(), index=True)
+    key = db.Column(db.String(255), index=True)
     index = db.Column(db.Integer())
     modtime = db.Column(db.DateTime(), server_default=db.func.now())
 
@@ -108,7 +108,7 @@ class EmailAttachment(db.Model):
 class GlobalObjectStorage(db.Model):
     __tablename__ = dbtableprefix + "globalobjectstorage"
     id = db.Column(db.Integer(), primary_key=True)
-    key = db.Column(db.Text(), index=True)
+    key = db.Column(db.String(1024), index=True)
     value = db.Column(db.Text())
     encrypted = db.Column(db.Boolean(), nullable=False, server_default=true())
     user_id = db.Column(db.Integer(), db.ForeignKey(dbtableprefix + 'user.id', ondelete='CASCADE'))