+++ docassemble_webapp/docassemble/webapp/listlog.py
@@ -10,7 +10,7 @@ ready_file = '/usr/share/docassemble/webapp/ready'
 
 @app.route('/listlog')
 def list_log_files():
-    result = subprocess.call("supervisorctl --serverurl http://localhost:9001 start sync > /dev/null && while supervisorctl --serverurl http://localhost:9001 status sync | grep -q RUNNING; do sleep 1; done", shell=True)
+    result = subprocess.run("supervisorctl --serverurl http://localhost:9001 start sync > /dev/null && while supervisorctl --serverurl http://localhost:9001 status sync | grep -q RUNNING; do sleep 1; done", shell=True).returncode
     if result == 0:
         return "\n".join(sorted([f for f in os.listdir(LOG_DIRECTORY) if os.path.isfile(os.path.join(LOG_DIRECTORY, f))]))
     else: