+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## [0.5.97] - 2020-02-17
+### Added
+- The `forget prior` option for `action buttons`.
+### Changed
+- Calls to subprocesses now have timeouts.
+### Fixed
+- Bug with table export.
+
 ## [0.5.96] - 2020-02-15
 ### Changed
 - `depends on` no longer implies `need`.